% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_temperature_rung.R
\name{add_temperature_rung}
\alias{add_temperature_rung}
\title{Add Rung to Temperature Ladder for Optimal Mixing in Parallel Tempering MCMC}
\usage{
add_temperature_rung(
  chain,
  target_imps,
  d,
  p,
  levels_dp,
  num_at_target,
  one_per_level = FALSE
)
}
\arguments{
\item{chain}{Matrix. The chain containing samples at the previous lowest level of a PTMCMC step (see \code{parallel_slice}). Contains \code{d} + \code{length(target_imps)} columns, where the first \code{d} entries in a row are the sample and the final entries are the target function (e.g. Implausibility) values at the different waves of the corresponding sample.}

\item{target_imps}{Numeric vector. Target implausibility levels for each chain.}

\item{d}{Integer. Dimensionality of the parameter space.}

\item{p}{Numeric. The quantile probability used to set the next level set (implausibility level).}

\item{levels_dp}{Integer. The number of decimal places for rounding the level set evaluation.}

\item{num_at_target}{Integer. The number of chains already fixed at the target level set values.}

\item{one_per_level}{Logical. If \code{TRUE}, ensures a unique chain for each target implausibility level. Set to \code{FALSE} for a more efficient search for the final target level in all waves and to \code{TRUE} if the goal is to have uniform samples for each wave for plotting or analysis. Defaults to \code{FALSE}.}
}
\value{
A list containing:
\describe{
\item{\code{new_imps}}{The adjusted implausibility levels.}
\item{\code{start_value}}{The starting value for the next iteration.}
\item{\code{num_at_target}}{The updated number of chains fixed at the target implausibility levels.}
}
}
\description{
This function adds a temperature/implausibility level to a current Parallel Tempering MCMC that has not yet warmed up to the target level.
}
\examples{
# Example usage:
set.seed(123)
chain <- matrix(runif(100), nrow = 10, ncol = 10)
target_imps <- c(0.1, 0.5, 1, 2, 5, Inf)
d <- 3
p <- 0.5
levels_dp <- 2
num_at_target <- 2
one_per_level <- TRUE
result <- add_temperature_rung(
  chain = chain,
  target_imps = target_imps,
  d = d,
  p = p,
  levels_dp = levels_dp,
  num_at_target = num_at_target,
  one_per_level = one_per_level
)
print(result)
}
