% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_slice.R
\name{one_slice}
\alias{one_slice}
\title{One Slice Sampling Step for a Single Chain for parallel tempering}
\usage{
one_slice(
  M,
  imp_level,
  final_levels,
  x_start,
  box_limits,
  final_num_waves,
  w = 1,
  implausibility,
  ...
)
}
\arguments{
\item{M}{Integer. The number of samples to draw.}

\item{imp_level}{Numeric vector. The implausibility levels for this chain.}

\item{final_levels}{Numeric vector. The final implausibility levels for all constraints.}

\item{x_start}{Numeric vector. The starting position for the chain.}

\item{box_limits}{Matrix. The box region where sampling is constrained; each row corresponds to a dimension.}

\item{final_num_waves}{Integer. The total number of implausibility constraints (waves).}

\item{w}{Numeric or numeric vector. The initial step size(s) for the slice sampler.}

\item{implausibility}{A user-defined function that calculates the implausibility
of a point given the required levels. (see main examples)}

\item{...}{Additional arguments to pass to the log-density function \code{f}.}
}
\value{
A list containing:
\describe{
\item{\code{Xt}}{Numeric vector. The final sampled position after the mutation step.}
\item{\code{Imp_t}}{Numeric vector. The implausibility values for the sampled position.}
}
}
\description{
This function performs slice sampling for a single chain within a parallel tempering algorithm using the \code{slice_sampler_core} function. (Often called mutation steps in the evolutionary Monte Carlo literature).
}
\note{
This function assumes that \code{f}, the log-density function, is defined globally or passed via \code{...}.
It also assumes that the implausibility function \code{implausibility} is globally defined and evaluates the constraints at the given position.
}
