% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_slice_all.R
\name{one_slice_all}
\alias{one_slice_all}
\title{One Slice All: Sampling from a Slice Sampler and Returning All Samples}
\usage{
one_slice_all(m, imp_level, final_levels, x_start,
 box_limits, final_num_waves, w = 1, implausibility)
}
\arguments{
\item{m}{Integer. The number of samples to generate.}

\item{imp_level}{Numeric vector. The current implausibility level for sampling.}

\item{final_levels}{Numeric vector. The final target implausibility levels.}

\item{x_start}{Numeric vector. The starting point for the slice sampler.}

\item{box_limits}{Numeric matrix. The limits of the sampling box.}

\item{final_num_waves}{Integer. The number of final waves to consider in the sampling process.}

\item{w}{Numeric. The width parameter used in the slice sampling process. Default is 1.}

\item{implausibility}{A user-defined function that calculates the implausibility
of a point given the required levels. (see main examples)}
}
\value{
A matrix where each row corresponds to a sample, and columns represent the sampled values along with
the corresponding implausibility levels.
}
\description{
This function performs slice sampling and returns all the samples along with their corresponding implausibility levels.
It uses \code{slice_sampler_core} for sampling and \code{Implausibility} to calculate implausibility values for each sample.
}
\examples{
# Example implausibility function
implausibility <- function(x, final_levels) {
  abs(x) # Placeholder: Euclidean norm as implausibility
}
result <- one_slice_all(
  m = 1000,
  imp_level = 0.5,
  final_levels = c(0.3, 0.7),
  x_start = c(0.2, 0.3),
  box_limits = matrix(c(0, 1, 0, 1), ncol = 2),
  final_num_waves = 2,
  implausibility=implausibility
)
}
