% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_slice.R
\name{parallel_slice}
\alias{parallel_slice}
\title{Parallel Tempering Slice Sampling for Uniform target regions}
\usage{
parallel_slice(
  num_chains,
  num_mutations,
  num_iterations,
  d,
  imp_levels,
  final_target_levels,
  x_starts,
  box_limits,
  debug_mode = FALSE,
  num_switches = 0,
  switch_settings_list = NULL,
  volume_ratio,
  print_every = 100,
  implausibility
)
}
\arguments{
\item{num_chains}{Integer. The number of parallel chains.}

\item{num_mutations}{Integer. The number of slice samples taken per chain per iteration (the term mutation arises from the Evolutionary Monte Carlo literature).}

\item{num_iterations}{Integer. The total number of iterations.}

\item{d}{Integer. The dimensionality of the parameter space.}

\item{imp_levels}{List of numeric vectors. The implausibility levels for each chain (excluding the first chain), see \code{note} below.}

\item{final_target_levels}{Numeric vector. The final implausibility levels for all constraints.}

\item{x_starts}{Matrix. \code{num_chains} x \code{d}. Starting positions for all chains except the first.}

\item{box_limits}{Matrix. The box region where sampling is constrained; each row corresponds to the lower and upper bounds of a dimension.}

\item{debug_mode}{Logical. If \code{TRUE}, runs in debug mode with \code{lapply}; otherwise uses \code{mclapply} for parallelism. MAY NEED TO CHANGE FOR WINDOWS}

\item{num_switches}{Integer. Number of switches for discrete settings in the first chain (default: \code{0}).}

\item{switch_settings_list}{List. A list of POSSIBLE settings for switches in the first chain.}

\item{volume_ratio}{Numeric. Ratio for scaling the step size \code{w} across chains.}

\item{print_every}{Integer. Frequency of status updates during iterations (default: \code{100}).}

\item{implausibility}{A user-defined function that calculates the implausibility
of a point given the required levels. (see main examples)}
}
\value{
A list containing:
\describe{
\item{\code{uniform_sample_list}}{A list of matrices, one per chain, containing the sampled values. The last chain contains samples from the \code{final_target_levels}. The matrices are \code{num_iterations}x \code{d+num_waves} where \code{num_waves} is the length of an element of \code{imp_levels}}.
\item{\code{restart}}{The state of the chains for restarting the sampler.}
}
}
\description{
This function performs parallel tempering MCMC with slice sampling for uniform samples from a level set of a function (called implausibility throughout).
}
\note{
The first chain samples uniformly over the box, so it does not use implausibility levels or starting values.
}
