% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sim_slice_partitioned.R
\name{partition_space}
\alias{partition_space}
\title{Partition the Sampling Space}
\usage{
partition_space(box_limits, n_partitions)
}
\arguments{
\item{box_limits}{Numeric matrix. A \verb{dims x 2} matrix where each row specifies the lower and upper bounds
of a dimension in the sampling space.}

\item{n_partitions}{Integer. The number of partitions to create along each dimension.}
}
\value{
A list where each element is a matrix representing a partition. Each matrix has two columns,
\code{"Lower"} and \code{"Upper"}, and rows corresponding to dimensions.
}
\description{
This function partitions a multidimensional box into smaller subspaces. Each partition is represented
as a set of lower and upper bounds for each dimension.
}
