% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler_core.R
\name{slice_sampler_core_once}
\alias{slice_sampler_core_once}
\title{Single Step of Uniform Slice Sampling}
\usage{
slice_sampler_core_once(x0, f, w, box_limits, max_retries = 100, ...)
}
\arguments{
\item{x0}{A numeric vector representing the current position in the parameter space.}

\item{f}{A function to evaluate the target distribution. The function should return -Inf for values outside the feasible region. -Inf is compared to a constant, but any f provided will lead to generation of uniform samples from the region where f > -Inf.}

\item{w}{A numeric vector or scalar of step sizes for each dimension.}

\item{box_limits}{A matrix with two columns representing the lower and upper bounds for each dimension of the box region.}

\item{max_retries}{The maximum number of retries to attempt if the sample fails to meet the constraints. Default is 100.}

\item{...}{Additional arguments passed to \code{f}.}
}
\value{
A list containing:
\itemize{
\item \code{sample}: The sampled point.
\item \code{w}: The adapted step sizes after the sampling step.
}
}
\description{
Performs a single step of uniform slice sampling. Given the current position \code{x0}, the function
samples from the target distribution by adaptively adjusting the bounds on each dimension.
}
\note{
If the sampler exceeds \code{max_retries}, a warning is issued, and the initial point \code{x0} is returned as the sample. This ensures continuity but may indicate a problem with the target density function or parameterization.
}
