% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydancer-package.R
\docType{package}
\name{tinydancer-package}
\alias{tinydancer}
\alias{tinydancer-package}
\title{tinydancer: Fast Uniform Samples of Tiny Subspaces for History Matching or Level Set Estimation}
\description{
This package helps to overcome a problem often encountered during history matching, but is possible during any kind of level set estimation, where the size of a target space/level set is small or tiny relative to the original parameter space. Standard approaches involve rejection sampling and often prove a major bottleneck in multi-wave history matching workflows. The methods in this package leverage parallel tempering and a custom subset simulation algorithm to generate uniform samples from tiny subspaces defined as level sets of a custom function called Implausibility. See examples and vignettes for more details.
}
\author{
\strong{Maintainer}: Daniel Williamson \email{d.williamson@exeter.ac.uk} (\href{https://orcid.org/0000-0001-8917-3300}{ORCID})

}
\keyword{internal}
